ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
pymupdf==1.23.3
python-docx==0.8.11
openai==1.2.4
langchain==0.0.335
pandas==2.1.1
spacy==3.7.2
sentence-transformers==2.2.2
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
